Refactor routes structure
