+++ app/routes/consumer/__init__.py
@@ -8,3 +8,5 @@ from . import cart  # noqa: E402
 from . import wallet  # noqa: E402
 from . import orders  # noqa: E402
 from . import shops  # noqa: E402
+from . import agent  # noqa: E402
+from . import items  # noqa: E402
