+++ app/routes/consumer/agent.py
@@ -1,15 +1,16 @@
 from flask import Blueprint, request, jsonify
-from app.utils import auth_required
+from app.utils import auth_required, role_required
 from app.version import API_PREFIX
-from agent.agent_core import run_agent
 
 # Optional blueprint for AI assistant
 agent_bp = Blueprint('agent', __name__, url_prefix=f'{API_PREFIX}/agent')
 
 @agent_bp.route('/query', methods=['POST'])
 @auth_required
+@role_required('consumer')
 def agent_query():
     """Handle AI assistant queries."""
+    from agent.agent_core import run_agent
     user_query = request.json.get('query') if request.is_json else None
     if not user_query:
         return jsonify({'status': 'error', 'message': 'Query is required'}), 400
