+++ app/routes/onboarding/__init__.py
@@ -1,0 +1,2 @@
+from . import auth  # noqa: F401
+from . import user  # noqa: F401
