+++ app/test_support.py
@@ -5,13 +5,15 @@ from app.utils import role_required
 import logging
 from app.services.wallet_ops import adjust_consumer_balance, adjust_vendor_balance, InsufficientFunds
 from app.utils import create_access_token, create_refresh_token
-from app.services.order_service import (
-    confirm_order as confirm_order_service,
-    confirm_modified_order as confirm_modified_order_service,
+from app.routes.consumer.orders import (
+    confirm_order_service,
+    confirm_modified_order_service,
     cancel_order_by_consumer as cancel_order_service,
+)
+from app.routes.vendor.orders import (
     update_status_by_vendor as update_status_service,
     cancel_order_by_vendor as cancel_order_vendor_service,
-    complete_return as complete_return_service,
+    service_complete_return as complete_return_service,
 )
 from models import db
 from models.user import UserProfile
